refactor of existing method checking

+++ data_ocean/converter.py
@@ -245,11 +245,8 @@ class Converter:
             i += 1
         if records_len:
             self.save_to_db(records)
-        if start_index == 0:
-            try:
-                self.delete_outdated()
-            except:
-                print('This converter has not ability to remove outdated records')
+        if start_index == 0 and hasattr(self.__class__, 'delete_outdated'):
+            self.delete_outdated()
         del elements
         print('All the records have been rewritten.')
